+++ awscli/customizations/s3/filegenerator.py
@@ -95,7 +95,7 @@ class FileDecodingError(Exception):
 class FileStat(object):
     def __init__(self, src, dest=None, compare_key=None, size=None,
                  last_update=None, src_type=None, dest_type=None,
-                 operation_name=None):
+                 operation_name=None, is_stream=False):
         self.src = src
         self.dest = dest
         self.compare_key = compare_key
@@ -104,6 +104,7 @@ class FileStat(object):
         self.src_type = src_type
         self.dest_type = dest_type
         self.operation_name = operation_name
+        self.is_stream = is_stream
 
 
 class FileGenerator(object):
@@ -115,7 +116,8 @@ class FileGenerator(object):
     ``FileInfo`` objects to send to a ``Comparator`` or ``S3Handler``.
     """
     def __init__(self, service, endpoint, operation_name,
-                 follow_symlinks=True, page_size=None, result_queue=None):
+                 follow_symlinks=True, page_size=None, result_queue=None,
+                 is_stream=False):
         self._service = service
         self._endpoint = endpoint
         self.operation_name = operation_name
@@ -124,6 +126,7 @@ class FileGenerator(object):
         self.result_queue = result_queue
         if not result_queue:
             self.result_queue = queue.Queue()
+        self.is_stream = is_stream
 
     def call(self, files):
         """
@@ -135,7 +138,11 @@ class FileGenerator(object):
         dest = files['dest']
         src_type = src['type']
         dest_type = dest['type']
-        function_table = {'s3': self.list_objects, 'local': self.list_files}
+        function_table = {'s3': self.list_objects}
+        if self.is_stream:
+            function_table['local'] = self.list_local_file_stream
+        else:
+            function_table['local'] = self.list_files
         sep_table = {'s3': '/', 'local': os.sep}
         source = src['path']
         file_list = function_table[src_type](source, files['dir_op'])
@@ -155,7 +162,15 @@ class FileGenerator(object):
                            compare_key=compare_key, size=size,
                            last_update=last_update, src_type=src_type,
                            dest_type=dest_type,
-                           operation_name=self.operation_name)
+                           operation_name=self.operation_name,
+                           is_stream=self.is_stream)
+
+    def list_local_file_stream(self, path, dir_op):
+        """
+        Yield some dummy values for a local file stream since it does not
+        actually have a file.
+        """
+        yield '-', 0, None
 
     def list_files(self, path, dir_op):
         """