+++ tests/unit/customizations/s3/__init__.py
@@ -16,7 +16,7 @@ from awscli.testutils import unittest
 import string
 
 import six
-from mock import patch
+from mock import patch, Mock
 
 
 class S3HandlerBaseTest(unittest.TestCase):
@@ -32,7 +32,6 @@ class S3HandlerBaseTest(unittest.TestCase):
 
     def tearDown(self):
         self.wait_timeout_patch.stop()
-
 
 def make_loc_files():
     """
@@ -161,6 +160,7 @@ def compare_files(self, result_file, ref_file):
     self.assertEqual(result_file.src_type, ref_file.src_type)
     self.assertEqual(result_file.dest_type, ref_file.dest_type)
     self.assertEqual(result_file.operation_name, ref_file.operation_name)
+    self.assertEqual(result_file.is_stream, ref_file.is_stream)
 
 
 def list_contents(bucket, session):
@@ -188,3 +188,24 @@ def list_buckets(session):
     html_response, response_data = operation.call(endpoint)
     contents = response_data['Buckets']
     return contents
+
+
+class MockStdIn(object):
+    """
+    This class patches stdin in order to write a stream of bytes into
+    stdin.
+    """
+    def __init__(self, input_bytes=b''):
+        input_data = six.BytesIO(input_bytes)
+        if six.PY3:
+            mock_object = Mock()
+            mock_object.buffer = input_data
+        else:
+            mock_object = input_data
+        self._patch = patch('sys.stdin', mock_object)
+
+    def __enter__(self):
+        self._patch.__enter__()
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        self._patch.__exit__()