+++ tests/unit/customizations/s3/test_subcommands.py
@@ -202,7 +202,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': local_file, 'dest': s3_file, 'filters': filters,
                   'paths_type': 'locals3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'cp', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -218,7 +219,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': local_file, 'dest': s3_file, 'filters': filters,
                   'paths_type': 'locals3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'cp', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -241,7 +243,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_file, 'dest': local_file, 'filters': filters,
                   'paths_type': 's3local', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'cp', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -258,7 +261,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_file, 'dest': s3_file, 'filters': filters,
                   'paths_type': 's3s3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'cp', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -275,7 +279,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_file, 'dest': s3_file, 'filters': filters,
                   'paths_type': 's3s3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'mv', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -292,7 +297,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_file, 'dest': s3_file, 'filters': filters,
                   'paths_type': 's3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'rm', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -313,7 +319,8 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': local_dir, 'dest': s3_prefix, 'filters': filters,
                   'paths_type': 'locals3', 'region': 'us-east-1',
                   'endpoint_url': None, 'verify_ssl': None,
-                  'follow_symlinks': True, 'page_size': None}
+                  'follow_symlinks': True, 'page_size': None,
+                  'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'sync', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -329,7 +336,7 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_prefix, 'dest': s3_prefix, 'paths_type': 's3',
                   'region': 'us-east-1', 'endpoint_url': None,
                   'verify_ssl': None, 'follow_symlinks': True,
-                  'page_size': None}
+                  'page_size': None, 'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'mb', params)
         cmd_arc.create_instructions()
         cmd_arc.run()
@@ -345,7 +352,7 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_prefix, 'dest': s3_prefix, 'paths_type': 's3',
                   'region': 'us-east-1', 'endpoint_url': None,
                   'verify_ssl': None, 'follow_symlinks': True,
-                  'page_size': None}
+                  'page_size': None, 'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'rb', params)
         cmd_arc.create_instructions()
         rc = cmd_arc.run()
@@ -362,7 +369,7 @@ class CommandArchitectureTest(S3HandlerBaseTest):
                   'src': s3_prefix, 'dest': s3_prefix, 'paths_type': 's3',
                   'region': 'us-east-1', 'endpoint_url': None,
                   'verify_ssl': None, 'follow_symlinks': True,
-                  'page_size': None}
+                  'page_size': None, 'is_stream': False}
         cmd_arc = CommandArchitecture(self.session, 'rb', params)
         cmd_arc.create_instructions()
         rc = cmd_arc.run()
@@ -472,6 +479,34 @@ class CommandParametersTest(unittest.TestCase):
         cmd_params = CommandParameters(self.session, 'rb', {'force': True},'')
         cmd_params.parameters['src'] = 's3://mybucket'
         cmd_params.check_force(None)
+
+    def test_validate_streaming_paths_upload(self):
+        parameters = {'src': '-', 'dest': 's3://bucket'}
+        cmd_params = CommandParameters(self.session, 'cp', parameters, '')
+        cmd_params._validate_streaming_paths()
+        self.assertTrue(cmd_params.parameters['is_stream'])
+        self.assertTrue(cmd_params.parameters['quiet'])
+        self.assertFalse(cmd_params.parameters['dir_op'])
+
+    def test_validate_streaming_paths_download(self):
+        parameters = {'src': 'localfile', 'dest': '-'}
+        cmd_params = CommandParameters(self.session, 'cp', parameters, '')
+        cmd_params._validate_streaming_paths()
+        self.assertTrue(cmd_params.parameters['is_stream'])
+        self.assertTrue(cmd_params.parameters['quiet'])
+        self.assertFalse(cmd_params.parameters['dir_op'])
+
+    def test_validate_no_streaming_paths(self):
+        parameters = {'src': 'localfile', 'dest': 's3://bucket'}
+        cmd_params = CommandParameters(self.session, 'cp', parameters, '')
+        cmd_params._validate_streaming_paths()
+        self.assertFalse(cmd_params.parameters['is_stream'])
+
+    def test_validate_streaming_paths_error(self):
+        parameters = {'src': '-', 'dest': 's3://bucket'}
+        cmd_params = CommandParameters(self.session, 'sync', parameters, '')
+        with self.assertRaises(ValueError):
+            cmd_params._validate_streaming_paths()
 
 
 class HelpDocTest(BaseAWSHelpOutputTest):